Switch back to internal classes.
