+++ spec/controllers/breeze/sitemap_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Breeze::SitemapController do
-
-  it "describes the api of SitemapController" 
-end
+# describe Breeze::SitemapController do
+# 
+#   it "describes the api of SitemapController" 
+# end
